+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.3] - 2023-02-11
+### Added
+- The `GetTransactions` method which returns all the transactions for an
+  account with optional query parameters.
+
 ## [0.0.2] - 2023-02-11
 ### Added 
 - The `Auth` method which returns the user's access token.